Initial Commit for Portfolio
I’m creating another repo for my portfolio that stands apart from any
work that I’ve done on it in previous classes. I intend to work on it
outside of any subsequent coursework. Eventually it will be
transitioned to a personal GitHub account, as soon as I understand how
to link the desktop app to two separate accounts!